functions/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.9.0
cryptography==46.0.3
deprecation==2.1.0
firebase_admin==7.1.0
firebase_functions==0.5.0
Flask==3.1.2
flask-cors==6.0.1
functions-framework==3.9.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-scheduler==2.17.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-events==0.5.0
google-genai>=0.1.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvicorn-worker==0.4.0
watchdog==6.0.0
Werkzeug==3.1.4
